+++ app/build.gradle
@@ -1,8 +1,19 @@
 apply plugin: 'com.android.application'
 apply plugin: 'kotlin-android'
 apply plugin: 'kotlin-android-extensions'
+apply plugin: 'kotlin-kapt'
 
 ext {
+    // app
+    app_versionCode=1
+    app_versionName="1.0"
+    app_applicationId="uk.co.sentinelweb.cuer.app"
+    app_compileSdkVersion=29
+    app_targetSdkVersion=29
+    app_minSdkVersion=21
+    app_buildToolsVersion="29.0.2"
+
+    // deps
     ver_core_ktx='1.2.0'
     ver_material='1.1.0'
     ver_constraintlayout='1.1.3'
@@ -14,24 +25,27 @@ ext {
     ver_mediarouter='1.1.0'
     ver_appcompat='1.1.0'
 
+    // deps test
     ver_junit='4.12'
     ver_mockk='1.9.3'
     ver_jfixture='2.7.2'
 
+    // deps androidTest
     ver_androidx_test_junit='1.1.1'
     ver_espresso_core='3.2.0'
+
 }
 
 android {
-    compileSdkVersion 29
-    buildToolsVersion "29.0.2"
+    compileSdkVersion app_compileSdkVersion
+    buildToolsVersion app_buildToolsVersion
 
     defaultConfig {
-        applicationId "uk.co.sentinelweb.cuer.app"
-        minSdkVersion 21
-        targetSdkVersion 29
-        versionCode 1
-        versionName "1.0"
+        applicationId app_applicationId
+        minSdkVersion app_minSdkVersion
+        targetSdkVersion app_targetSdkVersion
+        versionCode app_versionCode
+        versionName app_versionName
 
         testInstrumentationRunner "androidx.test.runner.AndroidJUnitRunner"
     }
@@ -61,6 +75,8 @@ android {
 }
 
 dependencies {
+    implementation project(":domain")
+    implementation project(":net")
     implementation fileTree(dir: 'libs', include: ['*.jar'])
     implementation "org.jetbrains.kotlin:kotlin-stdlib-jdk7:$ver_kotlin"
     implementation "androidx.appcompat:appcompat:$ver_appcompat"
@@ -78,13 +94,11 @@ dependencies {
     implementation  "org.koin:koin-android-viewmodel:$ver_koin"
 
     implementation "androidx.room:room-runtime:$ver_room"
-    annotationProcessor "androidx.room:room-compiler:$ver_room" // For Kotlin use kapt instead of annotationProcessor
+    kapt "androidx.room:room-compiler:$ver_room"
     // optional - Kotlin Extensions and Coroutines support for Room
     implementation "androidx.room:room-ktx:$ver_room"
     // optional - RxJava support for Room
-    implementation "androidx.room:room-rxjava2:$ver_room"
-    // optional - Guava support for Room, including Optional and ListenableFuture
-    implementation "androidx.room:room-guava:$ver_room"
+    // implementation "androidx.room:room-rxjava2:$ver_room"
     // Test helpers
     testImplementation "androidx.room:room-testing:$ver_room"
