+++ app/src/main/java/uk/co/sentinelweb/cuer/app/ui/browse/BrowseModel.kt
@@ -1,6 +1,6 @@
 package uk.co.sentinelweb.cuer.app.ui.browse
 
-import uk.co.sentinelweb.cuer.app.domain.MediaDomain
+import uk.co.sentinelweb.cuer.domain.MediaDomain
 
 data class BrowseModel constructor(
     val items:List<BrowseItemModel>
