+++ app/src/main/java/uk/co/sentinelweb/cuer/app/ui/player/PlayerModel.kt
@@ -1,6 +1,6 @@
 package uk.co.sentinelweb.cuer.app.ui.player
 
-import uk.co.sentinelweb.cuer.app.domain.MediaDomain
+import uk.co.sentinelweb.cuer.domain.MediaDomain
 
 data class PlayerModel constructor(
     val url: String,
