+++ domain/src/main/java/uk/co/sentinelweb/cuer/domain/PlaylistDomain.kt
@@ -1,4 +1,4 @@
-package uk.co.sentinelweb.cuer.app.domain
+package uk.co.sentinelweb.cuer.domain
 
 data class PlaylistDomain(
     val items: List<PlaylistItemDomain>
