+++ net/src/main/java/uk/co/sentinelweb/cuer/net/YoutubeDataInteractor.kt
@@ -1,0 +1,8 @@
+package uk.co.sentinelweb.cuer.net
+
+import uk.co.sentinelweb.cuer.domain.MediaDomain
+
+interface YoutubeDataInteractor {
+
+    suspend fun getInfo(id:String):MediaDomain
+}
